+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 It requires Dalli to be available in your dependencies at least 4.x.
 - The requirement for faraday 2+ has been relaxed, just faraday is now required in
 the gemspec. It has been tested with 2 and 1.
+- Redis session store has been updated so it only uses two keys
 
 ## [2.0.1] - 2026-03-27